Fix encoder correctness, improve drphylo aggregation, expand README
Encoder (encoder.hpp / encoder.cpp):
- Add skip_x param: treat 'X' like gap/unknown for protein/nucleotide
  datatypes (not universal); avoids inflating allele counts from
  unknown-AA symbols
- Add per-allele min_minor filter inside the column loop (matches old
  Python pipeline behaviour; was missing, so alleles below threshold
  were silently included)
- Change is_skip from static lambda to local array so skip_x can be
  set at runtime

train (main.cpp):
- Default min_minor raised from 1 to 2
- Pass skip_x to encoder based on --datatype flag
- Add --write-features / --write-features-transposed flags to dump
  the encoded feature matrix to CSV for debugging
- PSS output precision raised from 6 to 15 significant digits
- Median zero-filter threshold changed from 5e-7 to exactly 0.0 for
  gss/pss/bss median files

drphylo (main.cpp):
- Discover all lambda_N dirs produced by train and evaluate each
- Add --grid-rmse-cutoff and --grid-acc-cutoff thresholds to filter
  qualifying lambda models before aggregation
- Aggregate gene_predictions across qualifying models: predictions and
  intercepts are averaged; per-(gene,seq) scores use median of
  non-zero values; union of gene names is preserved
- Write aggregated eval_gene_predictions.txt, eval.txt, and
  eval_SPS_SPP.txt at the run level; generate SVG from aggregated data

Bug fixes (sg_lasso.cpp / sg_lasso_leastr.cpp):
- Fix memcpy size in altra(): sizeof(double) -> sizeof(float)
- Fix cross-validation subsetting: features.cols() -> features.rows()
- Remove redundant transpose: const arma::mat& A = features instead
  of A = features.t()

README.md: comprehensive rewrite with full usage reference

+++ src/sg_lasso.cpp
@@ -643,7 +643,7 @@ const arma::fcolvec SGLasso::altra(const arma::fcolvec& v_in,
         i=1;
     }
     else{
-        memcpy(x, v, sizeof(double) * n);
+        memcpy(x, v, sizeof(float) * n);
         i=0;
     }
 //std::cout << "Altra 3..." << std::endl;
